QLDA: fix UC22 notification cast and template-driven Excel exports
NguoiTrinhId is bigint in DB — map as long so CoreMessaging SP runs after approve.
Resolve entity type display by const value; keep PrintTemplates as layout SOT via HandMaintainedTemplate.

+++ QLDA.Domain/Entities/PheDuyet.cs
@@ -18,7 +18,10 @@ public class PheDuyet : Entity<Guid>, IAggregateRoot {
 
     public int? TrangThaiId { get; set; }
     public string? NoiDung { get; set; }
-    public int? NguoiTrinhId { get; set; }
+    /// <summary>
+    /// USER_MASTER.UserPortalId — DB column is bigint
+    /// </summary>
+    public long? NguoiTrinhId { get; set; }
     public DateTimeOffset NgayXuLy { get; set; }
 
     #region Navigation Properties
